Add message for users with no lists
Instead of just a blank screen, users with no todo lists will get a
helpful well with a message asking them to create a list, along with a
button linking to the new list form.

Also adds heading info consistent with that used in the todo_list#show
view, which will make more sense once user logins are implemented.