Turn data-confirm dialogs into bootstrap modals
Using the data-confirm-modal gem we turn rails data-confirm dialogs into
full bootstrap modals.

Include styling so we can set the class of the modal (defaulting it to
'.type-danger' and get nice fancy headers etc).

In order to custom style modals (and not use the default we set), this
relies on a custom version of the data-confirm-modal gem. A pull request
is in the works that will hopefully make it to the official version.

* Add the data-confirm-modal gem
* Load the gem's javascripts in application.js and set its custom
  defaults.
* Add custom styling for the modals
* Use the new confirmation verification to make users type in the name
  of a list before deleting it, if the list is not empty.