Add a welcome page for users who aren't logged in
Technically it's part of the TodoListController instead of the
StaticPagesController, because the root route is for
TodoListsController#index.

Updates styles to make jumbtrons on any controller able to push up and
directly touch the navbar (instead of affecting the navbar itself on
static pages)

Styling for outline buttons inside of jumbotrons fixed, now you can see
them.