WIP: TodoItems handle time zones in local time
Currently 'local time' is hard-coded as 'America/New_York' and server
time as UTC, but eventually users will set their own local time.

All forms and display convert the server/database time into local time.
Form inputs are done in local time, then converted back into UTC for
storage.

Form inputs have a 15-minute interval step to make things more
convenient... this may not be permanent.

The default Add Item is set to one day from the users current local
time, rounded down to the beginning of the hour.

List actions have been moved around to give a nice large check box on
the left-hand side. The title gets a strikethrough on completion and has
a small 'edit' link next to it.

Item editing form is implemented as a modal. Clear and reset buttons are
implemented for the date, particularly useful for iOS which otherwise
could not remove a due date once set. This is done in jQuery in reusable
way, so in theory clear and reset buttons could be added for any field.

Starting to move some time formatting helper methods out of the views,
although that is still a WIP. Need to add more specs to the views to
ensure all the time zone handling is done correctly.

This is a pretty big update but still rough around the edges, especially
where testing is concerned. Need to move some code around to helper
functions, or time-zone related stuff should probably be handled in the
user (so it knows about its own local time).

An alternative to asking the user for their local time zone is to use
GeoIP. Would that be better/worse?