Prevent widows in headings
Add a non-breaking space between the last two words in headings, so that
they don't end in an awkward single word when wrapped.

The .no-widows class can be added to any text element for this purpose,
but should always be added to the innermost element to work properly.
Otherwise, nested elements will be removed; only the text will remain.
  - e.g., if you have an 'h4.no-widows a' the link will be stripped, and
    the text inside it will become part of the h4