Flesh out missing tests for due dates
Using mutant, found areas in model code that where missing test
coverage, and wrote tests to get to near 100%.

For the time being, I'm not installing TimeCop to test the boundary
conditions of time comparisons (<= vs strictly <)

There seems to be undefined behavior in Rails? where `redirect_to` with
no parameters can redirect properly. I haven't come up with any tests
that will flush this out and, despite working in production, I'm
hesitant not to explicitly write out where the redirect is actually to,
lest the behavior act differently under some conditions.

Changed the status of redirects after PUT and DELETE to 303, because
AJAX was redirecting still under the PUT and DELETE verbs which could
lead to unintentional loss of data.
  - specifically, if I did an AJAX request to uncomplete an item, the
    DELETE request once completed, would redirect to the list itself
    still with the DELETE verb. Deleting the whole list. Oops.