todo_items/_show should have entire li
For when we transition to AJAX, we're going to want the entire list item
as part of the partial, to make it easier to send new/updated elements
as a response.

Stopped stubbing the todo_items/_show spec in the todo_list/show spec,
and instead added some additional stubbing for link_to calls from the
todo_items/_show_spec to the into todo_list/show_spec.