Add pry-rails and mutant-rspec to Gemfile
Officially adding these two gems to the dev environment, and making a
rake task for mutant to make it easier to run mutation tests.

Because of the time it takes and because some failure is OK, I'm not
adding mutant to CI at this time. Instead it's merely meant as a tool
to help decide where more tests are needed.