Hotfix: AJAX fixes for empty lists and archived items (#27)
* Archived items remain archived after AJAX calls

Due to duplicated ids, the AJAX calls weren't keeping archived items in
the archives tab. This is been fixed, along with specs added.

* Reload after creating first or destroying last itm

In order to ensure the messages for empty lists are correctly replaced
and readded, the AJAX responses force a reload when the first item is
created or the last item destroyed in a list.